[MIG] rotordc_optional_product: migration to 16.0
